Library for data containers

FlexibleSpacecraft.TimeLine.InitData — Type

struct InitiData

Struct that consists of the initial state value of the time-variant physical amounts in simulation

Base.getindex — Method

Base.getindex(v::Vector{<:SVector}, r::Int, datarow::Int)

get an element of the v<:SVector, used for custom data container for FlexibleSpacecraft.jl

Base.getindex — Method

Base.getindex(v::Vector{<:SVector}, r::AbstractRange, datarow::Int)

get a 1-D subset of the every datarow-th row of v::Vector{<:SVector} within r::AbstractRange, used for custom data container for FlexibleSpacecraft.jl

FlexibleSpacecraft.TimeLine.getdataindex — Method

function getdataindex(timerange::Tuple{<:Real, <:Real}, samplingtime::Real)::Union{UnitRange{Int64}, Colon}

returns an index::::Union{UnitRange{Int64}, Colon} that corresponding to the given timerange::Tuple{<:Real, <:Real}

FlexibleSpacecraft.TimeLine.initangularvelocity — Method

function initangularvelocity(simdata_num, initital_value::Vector)

Initialize array that contains time response of angular velocity

FlexibleSpacecraft.TimeLine.initquaterniondata — Method

function initquaterniondata(simdata_num, initial_value::Vector[4])

initialize array that contains time response of quaternion

FlexibleSpacecraft.TimeLine.initsimulationdata — Method

initsimulationdata(datanum::Int, initialdata::InitData)

Initialize the data container for the attitude dynamics