Plot recipe

FlexibleSpacecraft.jl have a post-processing features that allow you to create the plots of the simulation results.

FlexibleSpacecraft.PlotRecipe.FramePlot.dispframe — Method

function dispframe(time, refCoordinate, coordinate)

Generates the 3D figure of body fixed frame

FlexibleSpacecraft.PlotRecipe.FramePlot.framegif — Method

function frame_gif(time, Tsampling, refCoordinate, bodyCoordinateArray, Tgif = 0.4, FPS = 15)

Generates animation of frame rotation as GIF figure

FlexibleSpacecraft.PlotRecipe.PhysicalQuantity.angularvelocities — Method

function angularvelocities(time::StepRangeLen, angularvelocity::Vector{StaticArrays.SVector{3, <:Real}}; timerange::Tuple{<:Real, <:Real} = (0, 0))::AbstractPlot

Plots angular velocity of each axis in one figure

FlexibleSpacecraft.PlotRecipe.PhysicalQuantity.eulerangles — Method

eulerangles(time::StepRangeLen, eulerangle::Vector{StaticArrays.SVector{3, <:Real}}; timerange::Tuple{<:Real, <:Real} = (0, 0))::AbstractPlot

Plots time history of euler angles

FlexibleSpacecraft.PlotRecipe.PhysicalQuantity.quaternions — Method

function quaternions(time::StepRangeLen, quaternion::Vector{StaticArrays.SVector{4, <:Real}}; timerange::Tuple{<:Real, <:Real} = (0, 0))

Plot quaternions in single plot