Simulation API

This section contains information about the API for the simulation system of FlexibleSpacecraft.jl

Our API mainly contains the following sub-modules:

ParameterSetting: API for parameter setting for simulation core

The high-level API function for the simulation is runsimulation. Basically you can run your simulation by passing all the necessary arguments into function runsimulation.

ParameterSetting

FlexibleSpacecraft.ParameterSetting.SimulationConfig — Type

struct SimulationConfig

struct that contains the information about the simulation configuration

fields

simulationtime::Real: time length of the simulation

samplingtime::Real: sampling time of the simulation

FlexibleSpacecraft.ParameterSetting.setdisturbance — Method

setdisturbance(filepath::String)::DisturbanceConfig

set disturbance configuration from YAML setting file

FlexibleSpacecraft.ParameterSetting.setdynamicsmodel — Method

setdynamicsmodel(filepath::String)

Load the YAML file configuration and construct the appropriate model for the simulation

FlexibleSpacecraft.ParameterSetting.setinitvalue — Method

setinitvalue(filepath::String)::InitData

Define the inital value for simulation

FlexibleSpacecraft.ParameterSetting.setorbit — Method

setorbit(filepath::String, ECI::Frame)::OrbitInfo

Load the YAML file configuration and construct the appropriate model for the simulation

FlexibleSpacecraft.ParameterSetting.setsimconfig — Method

setsimconfig(filepath::String)::SimulationConfig

initialize the simulation configurations

Return value

simconfig::SimulationConfig