Visualization

FlexibleSpacecraft.jl have a post-processing features that allow you to create the plots of the simulation results.

FlexibleSpacecraft.VisualizationBase.FramePlot._Frame2Arrows — Method

_Frame2Arrows

FlexibleSpacecraft.VisualizationBase.FramePlot._plot_frame! — Method

_plot_frame!

FlexibleSpacecraft.VisualizationBase.FramePlot.animate_attitude — Method

frame_gif

Generates animation of frame rotation as GIF figure

FlexibleSpacecraft.VisualizationBase.AttitudeVisualization.plot_angular_momentum — Method

plot_angular_momentum

plot time history of angular momentum

FlexibleSpacecraft.VisualizationBase.AttitudeVisualization.plot_angularvelocity — Method

function plot_angularvelocity

Plots angular velocity of each axis in one figure

FlexibleSpacecraft.VisualizationBase.AttitudeVisualization.plot_eulerangles — Method

plot_eulerangles

Plots time history of euler angles

FlexibleSpacecraft.VisualizationBase.AttitudeVisualization.plot_quaternion — Method

plot_quaternion

Plot plot_quaternion in single plot