Disturbance input to attitude dynamics

Disturbance input for attitude motion is crucial to model how the spacecraft attitude changes from the effect of the outer environment. FlexibleSpacecraft.jl offers a disturbance input generation feature.

Disturbance input models

Constant torque input

Gravitational torque input

Parameter settings for disturbance input

You need an parameter setting YAML file to configure the disturbance input to the spacecraft. The file is presented as follows:

# set `property` distconfig to indicate that this file configures the disturbance input
property: distconfig

# set constant torque input with 3-d vector, define it to be [0, 0, 0] if no constant torque is applied
constant torque: [0.05, 0, 0]
# Boolean to select if the gravitational torque is considered or not.
gravitational torque: false

Libraries

FlexibleSpacecraft.Disturbance.DisturbanceConfig — Type

struct DisturbanceConfig

Struct that configures disturbance torque to the spacecraft

FlexibleSpacecraft.Disturbance.constant_torque — Method

constant_torque(constant_torque::Vector)

Function that returns given constant torque vector

FlexibleSpacecraft.Disturbance.constant_torque — Method

constant_torque(C_ECI2Body::SMatrix{3, 3, <:Real}, constant_torque::Nothing)

Function that returns zero vector

FlexibleSpacecraft.Disturbance.disturbanceinput — Method

FlexibleSpacecraft.Disturbance.gravity_gradient_torque — Method

function gravity_gradient_torque(inertia, orbit_angular_velocity, C_ECI2Body, C_ECI2LVLH, LVLHframe_z)

Function that returns gravity gradient torque

Arguments

Inertia matrix

Angular velocity of the orbit

Transfromation matrix from ECI frame to body frame

Transformation matrix from ECI frame to LVLH frame

Z-vector of LVLH frame