Dynamics simulation

Dynamics simulation is the key feature of the FlexibleSpacecraft.jl. This documentation provides the information on the features and APIs that allow you to execute your simulation conveniently.

DynamicsBase.jl module

DynamicsBase.jl is a submodule which is the basis of all the features on dynamics simulation with FlexibleSpacecraft.jl.

FlexibleSpacecraft.DynamicsBase.setdynamicsmodel — Method

setdynamicsmodel

Load a dictionaly data of configuration and construct the appropriate model for the simulation for the attitude dynamics

FlexibleSpacecraft.DynamicsBase.update_angularvelocity — Method

update_angularvelocity

update the angular velocity of the angular velocity of the attitude dynamics. Interface to the individual functions implemented in each submodules

Arguments

model: dynamics model for the atittude dynamics

currentTime::Real: current time

angularvelocity::AbstractVector{<:Real}: angular velocity vector

Tsampling::Real: sampling period for the dynamics simulation

currentbodyframe::Frame: current frame variable of the spacecraft's body fixed frame

distinput::AbstractVector{<:Real}: disturbance input vector

straccel::AbstractVector{<:Real}: acceleration of the structural response of the flexible appendages

strvelocity::AbstractVector{<:Real}: velocity of the structural response of the flexible appendages