Kinematics

KinematicsBase.jl module

KinematicsBase.jl is a submodule which is the basis of all the features on dynamics simulation with FlexibleSpacecraft.jl.

FlexibleSpacecraft.KinematicsBase.InitKinematicsData — Type

struct InitKinematicsData

Struct that consists of the initial state value of the time-variant physical amounts in simulation

FlexibleSpacecraft.KinematicsBase._calcdifferential_kinematics — Method

_calcdifferential_kinematics(omega::Vector, quaterion::Vector)

Get differential of quaternion from equation of kinematics

Arguments

omega: angular velocity of system

quaterion: current value of quaternion

Return

differential: differential of equation of kinematics

FlexibleSpacecraft.KinematicsBase._initangularvelocity — Method

function _initangularvelocity(simdata_num, initital_value::Vector)

Initialize array that contains time response of angular velocity

FlexibleSpacecraft.KinematicsBase._initquaternion — Method

function _initquaternion(simdata_num, initial_value::Vector[4])

initialize array that contains time response of quaternion

FlexibleSpacecraft.KinematicsBase.initattitudedata — Method

initattitudedata(datanum::Int, initialdata::InitKinematicsData)

Initialize the data container for the attitude dynamics

FlexibleSpacecraft.KinematicsBase.update_quaternion — Method

function update_quaternion(angularvelocity, quaternion, Tsampling)::SVector{4, <:Real}

calculate quaternion at next time step using 4th order Runge-Kutta method.