Core implementation of the simulation feature

This docs provides more detailed information on how FlexibleSpacecraft.jl executes simulation with provided submodules.

Simulation interface

Function runsimulation() is the top-most implementation of the simulation feature of FlexibleSpacecraft.jl.

FlexibleSpacecraft.SimulationCore.runsimulation — Function

runsimulation(attitudemodel, initvalue::Attitude.InitData, orbitinfo::Orbit.OrbitInfo, distconfig::DisturbanceConfig, simconfig::SimulationConfig)::Tuple

Function that runs simulation of the spacecraft attitude-structure coupling problem

Arguments

attitudemodel: Attitude dynamics model of the system

strmodel: Structural model of the flexible appendages

initvalue::InitData: Inital value of the simulation physical states

orbitinfo::Orbit.OrbitInfo: information and model definition of the orbital dynamics

distconfig::DisturbanceConfig: Disturbanve torque input configuration

simconfig::SimulationConfig: Simulation configuration ParameterSetting

Return value

Return is tuple of (time, attidata, orbitdata, strdata)

time: 1-D array of the time

attidata: Struct of time trajectory of the physical amount states of the spacecraft system

orbitdata: Struct of the orbit state trajectory

strdata: Struct of the trajectory of the state of the flexible appendage

Usage

(time, attitudedata, orbitdata, strdata) = runsimulation(attitudemodel, strmodel, initvalue, orbitinfo, distconfig, simconfig)