Simulation API

This section contains information about the API for the simulation system of FlexibleSpacecraft.jl

Our API mainly contains the following sub-modules:

ParameterSettingBase: API for parameter setting for simulation core

The high-level API function for the simulation is runsimulation. Basically you can run your simulation by passing all the necessary arguments into function runsimulation.

Parameter setting API

FlexibleSpacecraft.ParameterSettingBase.SimulationConfig — Type

struct SimulationConfig

struct that contains the information about the simulation configuration

fields

spacecraft::String: name of the spacecraft

note::String: note on the parameter settings

simulationtime::Real: time length of the simulation

samplingtime::Real: sampling time of the simulation

FlexibleSpacecraft.ParameterSettingBase._setinitvalue — Method

_setinitvalue(filepath::String)::InitData

Define the initial value for simulation

FlexibleSpacecraft.ParameterSettingBase._setsimconfig — Method

_setsimconfig(filepath::String)::SimulationConfig

initialize the simulation configurations

Return value

simconfig::SimulationConfig

FlexibleSpacecraft.ParameterSettingBase.readparamfile — Method

readparamfile

read the configuration of the spacecraft from the configuration file in YAML format.

Arguments

filepath::String: Path to the configuration file

Usage

paramfilepath = "./test/spacecraft2.yml"
(simconfig, attitudemodel, distconfig, initvalue, orbitinfo, strparam, strmodel) = readparamfile(paramfilepath)