Disturbance input to attitude dynamics

Disturbance input for attitude motion is crucial to model how the spacecraft attitude changes from the effect of the outer environment. A submodule AttitudeDisturbance.jl offers a disturbance input generation feature for FlexibleSpacecraft.jl.

Disturbance input models

Constant torque input

Step disturbance input trajectory

Gravitational torque input

Parameter settings for disturbance input

You need an parameter setting YAML file to configure the disturbance input to the spacecraft. The file is presented as follows:

disturbance:
    constant torque: [0.0, 0.0, 0.0]
    gravitational torque: false
    step trajectory:
        value: [[10, 0, 0], [0, 0, 0]]
        endurance: [1, 100]

Libraries for AttitudeDisturbance.jl

FlexibleSpacecraft.AttitudeDisturbance.DisturbanceConfig — Type

DisturbanceConfig

struct that configures disturbance torque to the spacecraft

Fields

consttorque::Union{SVector{3, <:Real}, Nothing}: constant torque input

gravitationaltorque::Bool: gravity gradient disturbance torque, configure with the boolean (true OR false)

steptraj::Union{Nothing, _ConfigStepTrajectory}: configuration for the step disturbance trajectory

FlexibleSpacecraft.AttitudeDisturbance.DisturbanceInternals — Type

DisturbanceInternals

internals of the disturbance torque for the attitude dynamics

Fields

steptraj::_InternalsStepTrajectory: internals of the step disturbance trajectory

FlexibleSpacecraft.AttitudeDisturbance._ConfigStepTrajectory — Type

_ConfigStepTrajectory

struct that configures the step trajectory of the disturbance torque input

FlexibleSpacecraft.AttitudeDisturbance._InternalsStepTrajectory — Type

_InternalsStepTrajectory

internals of the step trajectory

Fields

stepcnt::Unsigned: counter that specifies the current disturbance profile

duration::Real: duration time of the current disturbance profile

FlexibleSpacecraft.AttitudeDisturbance._constant_torque — Method

_constant_torque(torque_config::AbstractVector{<:Real})

Function that returns the predefined constant disturbance torque vector

FlexibleSpacecraft.AttitudeDisturbance._constant_torque — Method

constanttorque(torque_config::Nothing)

Function that returns zero disturbance torque for attitude dynamics

FlexibleSpacecraft.AttitudeDisturbance._gravity_gradient — Method

_gravity_gradient(inertia, orbit_angular_velocity, C_ECI2Body, C_ECI2LVLH, LVLHframe_z)

Function that returns gravity gradient torque

Arguments

Inertia matrix

Angular velocity of the orbit

Transfromation matrix from ECI frame to body frame

Transformation matrix from ECI frame to LVLH frame

Z-vector of LVLH frame

FlexibleSpacecraft.AttitudeDisturbance.calc_attitudedisturbance — Method

calc_attitudedisturbance

calculate disturbance torque input for the attitude dynamics

Arguments

distconfig::DisturbanceConfig: disturbance configuration

distinternals::DisturbanceInternals: disturbance internals

inertia::AbstractMatrix: inertia of the spacecraft body

currenttime::Real: current time

orbit_angular_velocity::Real: angular velocity of the orbital motion

C_ECI2Body::AbstractMatrix: rotation matrix from ECI to body

C_ECI2LVLH::AbstractMatrix: rotation matrix from ECI to LVLH

LVLHframe_z::AbstractVector: vector of the LVLH z-axix

Tsampling::Real: sampling time in simulation

FlexibleSpacecraft.AttitudeDisturbance.set_attitudedisturbance — Method

set_attitudedisturbance

set disturbance configuration from YAML setting file

YAML configuration format

disturbance:
    constant torque: {nothing OR "specify vector [0.0, 0.0, 0.0]"}
    gravitational torque: {false OR true}
    step trajectory:
        {
            nothing
        OR
            value: [[10, 0, 0], [0, 0, 0]]
            endurance: [1, 100]
        }