Kinematics

KinematicsBase.jl module

KinematicsBase.jl is a submodule which is the basis of all the features on dynamics simulation with FlexibleSpacecraft.jl.

FlexibleSpacecraft.KinematicsBase.InitData — Type

struct InitData

Struct that consists of the initial state value of the time-variant physical amounts in simulation

Base.Math.deg2rad — Method

Base.Math.deg2rad(rotationangle::Union{SVector{3, <:Real}, Vector{<:Real})

Convert rotation angle vector in degrees to radians

Base.Math.rad2deg — Method

Base.Math.rad2deg(rotationangle::Union{SVector{3, <:Real}, Vector{<:Real})

Convert rotation angle vector in radians to degrees

FlexibleSpacecraft.KinematicsBase._calcdifferential_kinematics — Method

_calcdifferential_kinematics(omega::Vector, quaterion::Vector)

Get differential of quaternion from equation of kinematics

Arguments

omega: angular velocity of system

quaterion: current value of quaternion

Return

differential: differential of equation of kinematics

FlexibleSpacecraft.KinematicsBase._initangularvelocity — Method

function _initangularvelocity(simdata_num, initital_value::Vector)

Initialize array that contains time response of angular velocity

FlexibleSpacecraft.KinematicsBase._initquaternion — Method

function _initquaternion(simdata_num, initial_value::Vector[4])

initialize array that contains time response of quaternion

FlexibleSpacecraft.KinematicsBase.dcm2euler — Method

dcm2euler(dcm::Union{SMatrix{3, 3, <:Real}, Matrix{<:Real}})::SVector{3, <:Real}

calculates z-y-x euler rotation angle from direction cosine matrix

FlexibleSpacecraft.KinematicsBase.dcm2quaternion — Method

dcm2quaternion(dcm::Matrix{Real})::Vector{Real}

calculate quaternion from direction cosine matrix (DCM) dcm

FlexibleSpacecraft.KinematicsBase.euler2dcm — Method

function eular2dcm(euler::Union{SVector{3, <:Real}, Vector{<:Real}})::SMatrix{3, 3, <:Real}

calculate direction cosine matrix from the vector of z-y-x eular angles.

Argument

euler::Union{SVector{3, <:Real}, Vector{<:Real}}: each element represents the rotation with z, y, x axis, respectively

FlexibleSpacecraft.KinematicsBase.euler2quaternion — Method

euler2quaternion(euler::Union{SVector{3, <:Real}, Vector{<:Real}})::SVector{4, Real}

calculates quaternion from z-y-x euler rotation angle

FlexibleSpacecraft.KinematicsBase.initattitudedata — Method

initattitudedata(datanum::Int, initialdata::InitData)

Initialize the data container for the attitude dynamics

FlexibleSpacecraft.KinematicsBase.quaternion2dcm — Method

quaternion2dcm(q::Union{Vector{<:Real}, SVector{4, <:Real}})

calculates direction cosine matrix from quaternion

FlexibleSpacecraft.KinematicsBase.quaternion2euler — Method

quaternion2euler(quaternion::Union{Vector{<:Real}, SVector{4, <:Real}})::SVector{3, <:Real}

calculates z-y-x euler rotation angle from quaternion

FlexibleSpacecraft.KinematicsBase.update_quaternion — Method

function update_quaternion(angularvelocity, quaternion, Tsampling)::SVector{4, <:Real}

calculate quaternion at next time step using 4th order Runge-Kutta method.