Rigid body dynamics model

Submodule RigidBody provides simulation algorithm for the simple rigid body spacecraft attitude dynamics. This model does not include the flexible structural appendages.

FlexibleSpacecraft.DynamicsBase.RigidBody — Module

module RigidBody

module that consists variables and functions needed for the simulation of rigid body spacecraft attitude dynamics. Internal module for FlexibleSpacecraft.jl

Usage

include("RigidBody.jl")
using .RigidBody

Parameter settings for rigid body attitude dynamics

Parameter setting YAML files should be organized as follows:

property: dynamics
name: example spacecraft
note: This model is for test purpose

dynamicsmodel: Rigid body # This must be set `Rigid body`

platform:
    inertia: [ # inertia matrix of the spacecraft
        50000, 0, 0,
        0, 50000, 0,
        0, 0, 50000
        ]

dynamicsmodel should be set Rigid body to tell the software that you are using the rigid body dynamics model.

Public interfaces

FlexibleSpacecraft.DynamicsBase.RigidBody.RigidBodyModel — Type

struct RigidBodyModel

Data container of rigid body spacecraft model. Used to specify and configure the parameter settings for simulation and control model in FlexibleSpacecraft.jl

Fields of struct RigidBodyModel

inertia::SMatrix{3, 3, <:Real}: Inertia matrix of spacecraft platform

FlexibleSpacecraft.DynamicsBase.update_angularvelocity — Function

update_angularvelocity

update the angular velocity of the angular velocity of the attitude dynamics. Interface to the individual functions implemented in each submodules

Arguments

model: dynamics model for the atittude dynamics

currentTime::Real: current time

angularvelocity::AbstractVector{<:Real}: angular velocity vector

Tsampling::Real: sampling period for the dynamics simulation

currentbodyframe::Frame: current frame variable of the spacecraft's body fixed frame

distinput::AbstractVector{<:Real}: disturbance input vector

straccel::AbstractVector{<:Real}: acceleration of the structural response of the flexible appendages

strvelocity::AbstractVector{<:Real}: velocity of the structural response of the flexible appendages